Refactoring of the code related to the actual HTML tags to avoid unnecessary repetitions
